Fixed signup user count, post-login Loader and welcome modal changes
